+++ app/src/views/private/components/license/use-license-guard.ts
